refactor(Controller): Modernize code
- Use getFirstNodeById for performance reason
- Add typing
- Use property contrustor promotion
- Use attributes instead of annotations

+++ lib/Controller/PhotosController.php
@@ -17,6 +17,8 @@ use OCA\Maps\Service\GeophotoService;
 use OCA\Maps\Service\PhotofilesService;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http;
+use OCP\AppFramework\Http\Attribute\NoAdminRequired;
+use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
 use OCP\AppFramework\Http\DataResponse;
 use OCP\DB\Exception;
 use OCP\Files\Folder;
@@ -27,52 +29,37 @@ use OCP\Files\NotPermittedException;
 use OCP\IRequest;
 
 class PhotosController extends Controller {
-	private $userId;
-	private $geophotoService;
-	private $photofilesService;
-	private $root;
-
-	public function __construct($AppName,
+	public function __construct(
+		string $appName,
 		IRequest $request,
-		GeophotoService $GeophotoService,
-		PhotofilesService $photofilesService,
-		IRootFolder $root,
-		$UserId) {
-		parent::__construct($AppName, $request);
-		$this->userId = $UserId;
-		$this->geophotoService = $GeophotoService;
-		$this->photofilesService = $photofilesService;
-		$this->root = $root;
+		private GeophotoService $geophotoService,
+		private PhotofilesService $photofilesService,
+		private IRootFolder $root,
+		private string $userId,
+	) {
+		parent::__construct($appName, $request);
 	}
 
 	/**
-	 * @NoAdminRequired
-	 * @NoCSRFRequired
-	 * @param null $myMapId
-	 * @param null $respectNoMediaAndNoimage
-	 * @param null $hideImagesOnCustomMaps
-	 * @param null $hideImagesInMapsFolder
-	 * @return DataResponse
 	 * @throws Exception
 	 * @throws NoUserException
 	 * @throws NotFoundException
 	 * @throws NotPermittedException
 	 */
-	public function getPhotos($myMapId = null, $respectNoMediaAndNoimage = null, $hideImagesOnCustomMaps = null, $hideImagesInMapsFolder = null): DataResponse {
+	#[NoAdminRequired]
+	#[NoCSRFRequired]
+	public function getPhotos(?int $myMapId = null, bool $respectNoMediaAndNoimage = true, bool $hideImagesOnCustomMaps = false, bool $hideImagesInMapsFolder = true): DataResponse {
 		$userFolder = $this->root->getUserFolder($this->userId);
-		if (is_null($myMapId) || $myMapId === '') {
-			$result = $this->geophotoService->getAll($this->userId, $userFolder, $respectNoMediaAndNoimage ?? true, $hideImagesOnCustomMaps ?? false, $hideImagesInMapsFolder ?? true);
+		if (is_null($myMapId)) {
+			$result = $this->geophotoService->getAll($this->userId, $userFolder, $respectNoMediaAndNoimage, $hideImagesOnCustomMaps, $hideImagesInMapsFolder);
 		} else {
-			$folders = $userFolder->getById($myMapId);
-			$folder = array_shift($folders);
-			$result = $this->geophotoService->getAll($this->userId, $folder, $respectNoMediaAndNoimage ?? true, $hideImagesOnCustomMaps ?? false, $hideImagesInMapsFolder ?? false);
+			$folder = $userFolder->getFirstNodeById($myMapId);
+			$result = $this->geophotoService->getAll($this->userId, $folder, $respectNoMediaAndNoimage, $hideImagesOnCustomMaps, $hideImagesInMapsFolder);
 		}
 		return new DataResponse($result);
 	}
 
 	/**
-	 * @NoAdminRequired
-	 * @NoCSRFRequired
 	 * @param int|null $myMapId
 	 * @param string|null $timezone
 	 * @param int $limit
@@ -86,13 +73,17 @@ class PhotosController extends Controller {
 	 * @throws NotFoundException
 	 * @throws NotPermittedException
 	 */
+	#[NoAdminRequired]
+	#[NoCSRFRequired]
 	public function getNonLocalizedPhotos(?int $myMapId = null, ?string $timezone = null, int $limit = 250, int $offset = 0, $respectNoMediaAndNoimage = null, $hideImagesOnCustomMaps = null, $hideImagesInMapsFolder = null): DataResponse {
 		$userFolder = $this->root->getUserFolder($this->userId);
-		if (is_null($myMapId) || $myMapId === '') {
+		if (is_null($myMapId)) {
 			$result = $this->geophotoService->getNonLocalized($this->userId, $userFolder, $respectNoMediaAndNoimage ?? true, $hideImagesOnCustomMaps ?? false, $hideImagesInMapsFolder ?? true, $timezone, $limit, $offset);
 		} else {
-			$folders = $userFolder->getById($myMapId);
-			$folder = array_shift($folders);
+			$folder = $userFolder->getFirstNodeById($myMapId);
+			if ($folder instanceof Folder) {
+				throw new NotFoundException('Could find map with mapid: ' . $myMapId);
+			}
 			$result = $this->geophotoService->getNonLocalized($this->userId, $folder, $respectNoMediaAndNoimage ?? true, $hideImagesOnCustomMaps ?? false, $hideImagesInMapsFolder ?? false, $timezone, $limit, $offset);
 		}
 		return new DataResponse($result);
@@ -100,12 +91,9 @@ class PhotosController extends Controller {
 
 
 	/**
-	 * @NoAdminRequired
 	 * @param $paths
 	 * @param $lats
 	 * @param $lngs
-	 * @param bool $directory
-	 * @param null $myMapId
 	 * @param bool $relative
 	 * @return DataResponse
 	 * @throws NoUserException
@@ -113,10 +101,11 @@ class PhotosController extends Controller {
 	 * @throws NotPermittedException
 	 * @throws InvalidPathException
 	 */
-	public function placePhotos($paths, $lats, $lngs, bool $directory = false, $myMapId = null, bool $relative = false): DataResponse {
+	#[NoAdminRequired]
+	public function placePhotos($paths, $lats, $lngs, bool $directory = false, ?int $myMapId = null, bool $relative = false): DataResponse {
 		$userFolder = $this->root->getUserFolder($this->userId);
-		if (!is_null($myMapId) and $myMapId !== '') {
-			if ($directory === 'true') {
+		if (!is_null($myMapId)) {
+			if ($directory) {
 				// forbid folder placement in my-maps
 				throw new NotPermittedException();
 			}
@@ -137,7 +126,7 @@ class PhotosController extends Controller {
 				foreach ($paths as $key => $path) {
 					$photoFile = $userFolder->get($path);
 					// is the photo in this map's folder?
-					if (!$folder->getById($photoFile->getId())) {
+					if ($folder->getFirstNodeById($photoFile->getId()) !== null) {
 						$copiedFile = $photoFile->copy($folder->getPath() . '/' . $photoFile->getName());
 						$paths[$key] = $userFolder->getRelativePath($copiedFile->getPath());
 					}
@@ -149,22 +138,20 @@ class PhotosController extends Controller {
 	}
 
 	/**
-	 * @NoAdminRequired
 	 * @param $paths
-	 * @return DataResponse
 	 */
-	public function resetPhotosCoords($paths, $myMapId = null): DataResponse {
+	#[NoAdminRequired]
+	public function resetPhotosCoords($paths, ?int $myMapId = null): DataResponse {
 		$userFolder = $this->root->getUserFolder($this->userId);
 		$result = [];
 		if (sizeof($paths) > 0) {
 			$result = $this->photofilesService->resetPhotosFilesCoords($this->userId, $paths);
 		}
-		if (!is_null($myMapId) and $myMapId !== '') {
+		if (!is_null($myMapId)) {
 			foreach ($paths as $key => $path) {
-				$folders = $userFolder->getById($myMapId);
-				$folder = array_shift($folders);
+				$folder = $userFolder->getFirstNodeById($myMapId);
 				$photoFile = $userFolder->get($path);
-				if ($folder->isSubNode($photoFile)) {
+				if ($folder instanceof Folder && $folder->isSubNode($photoFile)) {
 					$photoFile->delete();
 					unset($paths[$key]);
 				}
@@ -173,10 +160,7 @@ class PhotosController extends Controller {
 		return new DataResponse($result);
 	}
 
-	/**
-	 * @NoAdminRequired
-	 * @return DataResponse
-	 */
+	#[NoAdminRequired]
 	public function clearCache(): DataResponse {
 		$result = $this->geophotoService->clearCache();
 		if ($result) {
@@ -186,10 +170,7 @@ class PhotosController extends Controller {
 		}
 	}
 
-	/**
-	 * @NoAdminRequired
-	 * @return DataResponse
-	 */
+	#[NoAdminRequired]
 	public function getBackgroundJobStatus(): DataResponse {
 		return new DataResponse($this->photofilesService->getBackgroundJobStatus($this->userId));
 	}
